Revert "moved DirectusProvider into base folder"
This reverts commit 3fc5658fb1a0ef02e41d2da47dbaf7b8566787e4.
