+++ tsconfig.json
@@ -15,7 +15,6 @@
     "paths": {
       "@/*": ["*"],
       "@components/*": ["components/*"],
-      "@base/*": ["base/*"],
     }
   }
 }
